refactor(models): split models into relevant files and add mapping datatype to models
